bootstrap in, toolbar works, footer in
unable to implement responsive images.  it is totally counter-intuitive that on a bootstrap-oriented web page i was supposed to know that I had to write additional custom CSS code to implement such a feature.  put in some rows & columns but unable to make any of the bootstrap classes alleged to handle images actually work in practice.  Again, it seems I was supposed to write iterations of custom CSS --on a bootstrap-driven web page?  in order to ...shrink images? I am EXTREMELY DISSATISFIED by the apparent gap between the written instructions and the unwritten expectations for this seemingly-simple task.